perf(compose): Optimize SeaweedFS docker compose startup/shutdown
The SeaweedFS deployment is put in a separate compose file and each
component is started separately with a shared data volume. This makes it
more transparent which component fails and allows us to use different
RAFT parameters which significantly lowers the master election time at
startup (for some reason this does not work in monolith mode).

Signed-off-by: Alf-Rune Siqveland <alf.rune@siqveland.no>